Move openai.py to beyond_the_loop module
